Organize stuff a little more
